requirements.txt
streamlit==1.26.0
google-generativeai==0.5.0
matplotlib==3.8.0
seaborn==0.12.2
pandas==2.1.1